pyproject.toml
[project]
name = "air-lab-os"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["runtime", "lab", "evaluation", "memory", "patterns", "datasets"]

[tool.pytest.ini_options]
pythonpath = ["."]
